Attempted to use multiple types of compute functions
The two types of compute functions I want to accept are:
- Takes an aggregator argument: (V, Option[C], Option[A], Int) => (V, Iterable[M])
- Does not take an aggregator argument: (V, Option[C], Int) => (V, Iterable[M])

This commit attempts to do this by defining an implicit conversion, addAggregatorArg, which converts from the second to the first. However, Scala isn't able to infer the type parameter A of run properly; it should be Nothing in this case, but that doesn't get inferred.