Switch from pmap to fcl for building web graph
The WebGraph project stores the node-URL maps in two different
formats: as an external prefix map or as a front-coded list. The
former uses an on-disk structure and so avoids using too much
memory. The latter stores everything in-memory and so is much faster.

The external prefix map turned out to be too slow (would have taken
~11 days to parse a 100 million node graph using an m1.large EC2
instance), so this commit switches to a front-coded list, which did it
in ~3 hours.